+++ src/lib/layout/Sidebar.js
@@ -63,20 +63,20 @@ export default class Sidebar extends Component {
     this.setComponentTop()
   }
 
-  renderGroupContent (group, groupTitleKey) {
+  renderGroupContent (group, isRightSidebar, groupTitleKey, groupRightSidebarKey) {
     if (this.props.groupRenderer) {
-      return React.createElement(this.props.groupRenderer, { group })
+      return React.createElement(this.props.groupRenderer, { group, isRightSidebar })
     } else {
-      return _get(group, groupTitleKey)
+      return _get(group, isRightSidebar ? groupRightSidebarKey : groupTitleKey)
     }
   }
 
   render () {
     const {
-      fixedHeader, width, lineHeight, zIndex, groupHeights, height, headerHeight
+      fixedHeader, width, lineHeight, zIndex, groupHeights, height, headerHeight, isRightSidebar
     } = this.props
 
-    const {groupIdKey, groupTitleKey} = this.props.keys
+    const {groupIdKey, groupTitleKey, groupRightSidebarKey} = this.props.keys
 
     const {
       scrollTop
@@ -126,14 +126,14 @@ export default class Sidebar extends Component {
 
       groupLines.push(
         <div key={_get(group, groupIdKey)} className={'rct-sidebar-row' + (i % 2 === 0 ? ' rct-sidebar-row-even' : ' rct-sidebar-row-odd')} style={elementStyle}>
-          {this.renderGroupContent(group, groupTitleKey)}
+          {this.renderGroupContent(group, isRightSidebar, groupTitleKey, groupRightSidebarKey)}
         </div>
       )
       i += 1
     })
 
     return (
-      <div ref='sidebar' className='rct-sidebar' style={sidebarStyle}>
+      <div ref='sidebar' className={'rct-sidebar' + (isRightSidebar ? ' rct-sidebar-right' : '')} style={sidebarStyle}>
         {header}
         <div style={groupsStyle}>
           {groupLines}
@@ -151,10 +151,12 @@ Sidebar.propTypes = {
   fixedHeader: React.PropTypes.oneOf(['fixed', 'absolute', 'none']),
   keys: React.PropTypes.object.isRequired,
   groupRenderer: React.PropTypes.func,
-  children: React.PropTypes.node
+  children: React.PropTypes.node,
+  isRightSidebar: React.PropTypes.bool
 }
 Sidebar.defaultProps = {
   fixedHeader: 'none',
   zIndex: 12,
-  children: null
+  children: null,
+  isRightSidebar: false
 }